eth-accounts module simplified and TransactionSigner moved to eth module
